Update: Finished shop products CRUD

+++ src/app/shop/components/shoplist.tsx
@@ -4,13 +4,16 @@ import Link from 'next/link';
 import React, { useEffect, useState } from 'react';
 import { ShopService } from "../lib/service";
 import { Shop } from "../lib/model";
-import { BiTrash } from 'react-icons/bi';
+import { BiTrash, BiEdit } from 'react-icons/bi';
+import UpdateForm from './updateForm';
 
 const shopService = new ShopService();
 
 const ShopList: React.FC = () => {
     const [shops, setShops] = useState<Shop[]>([]);
     const [error, setError] = useState<string | null>(null);
+    const [selectedShop, setSelectedShop] = useState<Shop | null>(null);
+    const [isPopupOpen, setIsPopupOpen] = useState<boolean>(false);
 
     const fetchShops = async () => {
         try {
@@ -27,17 +30,25 @@ const ShopList: React.FC = () => {
     };
 
     const deleteShop = async (id: string) => {
-        try {
-            await shopService.deleteShop(id);
+        if (window.confirm("¿Está seguro de que desea eliminar este negocio?")) {
             setShops(shops.filter(shop => shop.id !== id));
-        } catch (error) {
-            if (error instanceof Error) {
-                setError(error.message);
-            } else {
-                setError("Ha ocurrido un error desconocido");
-            }
-            console.error("Error eliminando negocio:", error);
+            await shopService.deleteShop(id);
         }
+    };
+
+    const openPopup = (shop: Shop) => {
+        setSelectedShop(shop);
+        setIsPopupOpen(true);
+    };
+
+    const closePopup = () => {
+        setSelectedShop(null);
+        setIsPopupOpen(false);
+    };
+
+    const handleUpdate = async () => {
+        await fetchShops();
+        closePopup();
     };
 
     useEffect(() => {
@@ -58,12 +69,22 @@ const ShopList: React.FC = () => {
                             <button onClick={() => deleteShop(shop.id)}>
                                 <BiTrash />
                             </button>
-
+                            <button onClick={() => openPopup(shop)}>
+                                <BiEdit />
+                            </button>
                         </li>
                     ))}
                 </ul>
             ) : (
                 <p>No hay negocios disponibles.</p>
+            )}
+
+            {isPopupOpen && selectedShop && (
+                <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50">
+                    <div className="bg-white p-4 rounded-md">
+                        <UpdateForm shop={selectedShop} onClose={closePopup} onUpdate={handleUpdate} />
+                    </div>
+                </div>
             )}
         </>
     );
